fix/improvement: CheckoutConfigService location and export change
