+++ projects/storefrontlib/src/cms-components/checkout/checkout.module.ts
@@ -1,6 +1,5 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
-import { CheckoutConfigService } from './checkout-config.service';
 import { CheckoutOrchestratorModule } from './components/checkout-orchestrator/checkout-orchestrator.module';
 import { CheckoutOrderSummaryModule } from './components/checkout-order-summary/checkout-order-summary.module';
 // tslint:disable-next-line
@@ -30,7 +29,5 @@ import { ShippingAddressModule } from './components/shipping-address/shipping-ad
     ReviewSubmitModule,
     ShippingAddressModule,
   ],
-  // @todo: should we keep below provider here?
-  providers: [CheckoutConfigService],
 })
 export class CheckoutComponentModule {}
