+++ projects/storefrontlib/src/cms-components/checkout/components/delivery-mode/delivery-mode.component.spec.ts
@@ -10,7 +10,7 @@ import {
   RoutingService,
 } from '@spartacus/core';
 import { Observable, of } from 'rxjs';
-import { CheckoutConfigService } from '../../checkout-config.service';
+import { CheckoutConfigService } from '../../services/checkout-config.service';
 import { DeliveryModeComponent } from './delivery-mode.component';
 
 import createSpy = jasmine.createSpy;
