+++ projects/storefrontlib/src/cms-components/checkout/components/delivery-mode/delivery-mode.component.ts
@@ -13,7 +13,7 @@ import {
 } from '@spartacus/core';
 import { Observable, Subscription } from 'rxjs';
 import { map } from 'rxjs/operators';
-import { CheckoutConfigService } from '../../checkout-config.service';
+import { CheckoutConfigService } from '../../services/checkout-config.service';
 
 @Component({
   selector: 'cx-delivery-mode',
