+++ projects/storefrontlib/src/cms-components/checkout/guards/delivery-mode-set.guard.spec.ts
@@ -4,9 +4,9 @@ import { RouterTestingModule } from '@angular/router/testing';
 import { RoutesConfig, RoutingConfigService } from '@spartacus/core';
 import { Observable, of } from 'rxjs';
 import { defaultStorefrontRoutesConfig } from '../../../cms-structure/routing/default-routing-config';
-import { CheckoutConfigService } from '../checkout-config.service';
 import { CheckoutConfig } from '../config/checkout-config';
 import { defaultCheckoutConfig } from '../config/default-checkout-config';
+import { CheckoutConfigService } from '../services/checkout-config.service';
 import { CheckoutDetailsService } from '../services/checkout-details.service';
 import { DeliveryModeSetGuard } from './delivery-mode-set.guard';
