+++ projects/storefrontlib/src/cms-components/checkout/services/index.ts
@@ -1,1 +1,2 @@
+export * from './checkout-config.service';
 export * from './checkout-details.service';
